refactor(multipooler): extract a pgBackRest engine package from manager
Move all pgBackRest interaction into a new internal/manager/backup package
(backup.Engine): backup, restore, expire, verify, list, check, and
archive-command configuration. The manager keeps the gRPC handlers, the
restore/first-backup orchestration, policy, and PG lifecycle, delegating the
pgBackRest steps to the engine.

- Backup/restore metrics move into the package (backup.Metrics), owned by the
  engine; the manager records lock-wait via Engine.Metrics().
- The engine depends only on a small Identity interface (Id/ShardKey), an
  injected command runner, and config that is settable at runtime (pgbackrest
  config path + pgpass resolved at topology load, repo config at DB-open). It
  has no role or consensus dependency.
- Role/consensus-driven decisions stay in the manager: allowBackupOnPrimary and
  getPrimaryAsPg2Args own the policy and backup-source resolution and pass the
  results into the engine, so the engine never reads PoolerType.
- Backup type is validated and mapped to a typed backup.PgBackRestType before
  the source is resolved, so an invalid type is reported up front.
