+++ go/services/multipooler/internal/manager/backup/backup.go
@@ -1,0 +1,205 @@
+// Copyright 2026 Supabase, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package backup
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	commonbackup "github.com/multigres/multigres/go/common/backup"
+	"github.com/multigres/multigres/go/common/eventlog"
+	"github.com/multigres/multigres/go/common/mterrors"
+	"github.com/multigres/multigres/go/common/topoclient"
+	clustermetadatapb "github.com/multigres/multigres/go/pb/clustermetadata"
+	mtrpcpb "github.com/multigres/multigres/go/pb/mtrpc"
+	"github.com/multigres/multigres/go/services/multipooler/internal/manager/actionlock"
+	"github.com/multigres/multigres/go/tools/telemetry"
+)
+
+// Backup performs the backup steps inside the parent "backup" span.
+// Caller must hold both the action lock and the backup lease.
+func (e *Engine) Backup(ctx context.Context, pgBackRestType PgBackRestType, jobID string, pg2Args []string, poolerType clustermetadatapb.PoolerType) (retBackupID string, retErr error) {
+	if err := actionlock.AssertActionLockHeld(ctx); err != nil {
+		return "", err
+	}
+	if err := topoclient.AssertBackupLockHeld(ctx, e.id.ShardKey()); err != nil {
+		return "", mterrors.Wrap(err, "backup lease not held")
+	}
+
+	cfg, err := e.requireBackupConfig()
+	if err != nil {
+		return "", err
+	}
+
+	e.metrics.IncBackupAttempts(ctx)
+	defer func() {
+		if retErr == nil {
+			e.metrics.IncBackupSuccesses(ctx)
+		} else {
+			e.metrics.IncBackupFailures(ctx)
+		}
+	}()
+
+	e.logger.InfoContext(ctx, "Starting backup operation", "backup_type", cfg.Type(), "pgbackrest_type", pgBackRestType)
+
+	// Get path to pgbackrest.conf generated by pgctld
+	configPath, err := e.requireConfigPath()
+	if err != nil {
+		return "", err
+	}
+
+	tableGroup := e.id.ShardKey().GetTableGroup()
+	shard := e.id.ShardKey().GetShard()
+	multipoolerID := topoclient.MultiPoolerIDString(e.id.Id())
+	multipoolerName := e.id.Id().Name
+
+	// Use provided job_id or generate one (same format as multiadmin)
+	effectiveJobID := jobID
+	if effectiveJobID == "" {
+		effectiveJobID = commonbackup.GenerateJobID(multipoolerID)
+	}
+
+	// Emit Started now that we have a correlation ID and parameters are valid.
+	// BackupName uses effectiveJobID since the actual pgbackrest label isn't assigned yet.
+	eventlog.Emit(ctx, e.logger, eventlog.Started, eventlog.BackupAttempt{BackupName: effectiveJobID})
+	defer func() {
+		if retErr == nil {
+			// retBackupID is set to the actual pgbackrest label by the success return.
+			eventlog.Emit(ctx, e.logger, eventlog.Success, eventlog.BackupAttempt{BackupName: retBackupID})
+		} else {
+			eventlog.Emit(ctx, e.logger, eventlog.Failed, eventlog.BackupAttempt{BackupName: effectiveJobID}, "error", retErr)
+		}
+	}()
+
+	// Execute pgbackrest backup command
+	ctx, cancel := context.WithTimeout(ctx, commonbackup.BackupTimeout)
+	defer cancel()
+
+	args := []string{
+		"--stanza=" + stanzaName,
+		"--config=" + configPath,
+		"--type=" + string(pgBackRestType),
+	}
+
+	// pg2 args for replica backups (empty for primary) are resolved by the
+	// manager (which owns role + the consensus-recorded primary) and passed in.
+	args = append(args, pg2Args...)
+
+	// Reject backups with empty table_group or shard to prevent corrupt metadata.
+	// Backups missing these annotations would be silently skipped by replicas
+	// during restore, causing them to fall back to older backups.
+	if tableGroup == "" {
+		return "", mterrors.New(mtrpcpb.Code_FAILED_PRECONDITION, "table_group is missing")
+	}
+	if shard == "" {
+		return "", mterrors.New(mtrpcpb.Code_FAILED_PRECONDITION, "shard is missing")
+	}
+	args = append(args, "--annotation=table_group="+tableGroup)
+	args = append(args, "--annotation=shard="+shard)
+
+	// Add multipooler_id, pooler_type, and job_id annotations for unique identification
+	args = append(args, "--annotation=multipooler_id="+multipoolerName)
+	args = append(args, "--annotation=pooler_type="+poolerType.String())
+	args = append(args, "--annotation=job_id="+effectiveJobID)
+
+	args = append(args, "backup")
+
+	cmd := e.pgbackrestCmd(ctx, args...)
+
+	// Execute backup with progress logging
+	var output []byte
+	backupStart := time.Now()
+	err = telemetry.WithSpan(ctx, "backup/pgbackrest", func(ctx context.Context) error {
+		var runErr error
+		output, runErr = e.run(ctx, cmd, "pgbackrest backup")
+		return runErr
+	})
+	e.metrics.RecordBackupDuration(ctx, time.Since(backupStart).Seconds())
+	if err != nil {
+		return "", mterrors.New(mtrpcpb.Code_INTERNAL,
+			fmt.Sprintf("pgbackrest backup failed: %v\nOutput: %s", err, string(output)))
+	}
+
+	// Find the backup ID by querying pgbackrest info with our unique annotations
+	var foundBackupID string
+	err = telemetry.WithSpan(ctx, "backup/find", func(ctx context.Context) error {
+		var findErr error
+		foundBackupID, findErr = e.FindByJobID(ctx, effectiveJobID)
+		return findErr
+	})
+	if err != nil {
+		return "", mterrors.New(mtrpcpb.Code_INTERNAL,
+			fmt.Sprintf("failed to find backup by annotations: %v\nBackup output: %s", err, string(output)))
+	}
+
+	// Verify the backup to ensure it's valid
+	verifyCtx, verifyCancel := context.WithTimeout(ctx, commonbackup.VerifyTimeout)
+	defer verifyCancel()
+
+	verifyCmd := e.pgbackrestCmd(verifyCtx,
+		"--stanza="+stanzaName,
+		"--config="+configPath,
+		"--set="+foundBackupID,
+		"verify")
+
+	// Execute verify with progress logging
+	var verifyOutput []byte
+	verifyStart := time.Now()
+	verifyErr := telemetry.WithSpan(verifyCtx, "backup/verify", func(ctx context.Context) error {
+		var runErr error
+		verifyOutput, runErr = e.run(ctx, verifyCmd, "pgbackrest verify backup_id="+foundBackupID)
+		return runErr
+	})
+	e.metrics.RecordBackupVerifyDuration(ctx, time.Since(verifyStart).Seconds())
+	if verifyErr != nil {
+		return "", mterrors.New(mtrpcpb.Code_INTERNAL,
+			fmt.Sprintf("pgbackrest verify failed for backup %s: %v\nOutput: %s", foundBackupID, verifyErr, string(verifyOutput)))
+	}
+	// TODO: use `pgbackrest info` to verify that the database pages from the backed up
+	// Postgres cluster pass checksum validation.
+
+	return foundBackupID, nil
+}
+
+// PgBackRestType is a validated pgBackRest backup type — the value passed to
+// pgbackrest's --type flag. Construct via ValidateBackupType.
+type PgBackRestType string
+
+const (
+	PgBackRestFull PgBackRestType = "full"
+	PgBackRestDiff PgBackRestType = "diff"
+	PgBackRestIncr PgBackRestType = "incr"
+)
+
+// ValidateBackupType validates the requested backup type and maps it to the
+// pgBackRest type. The manager calls this before resolving the backup source so
+// a bad type is rejected up front.
+func ValidateBackupType(backupType string) (PgBackRestType, error) {
+	if backupType == "" {
+		return "", mterrors.New(mtrpcpb.Code_INVALID_ARGUMENT, "type is required")
+	}
+	switch backupType {
+	case "full":
+		return PgBackRestFull, nil
+	case "differential":
+		return PgBackRestDiff, nil
+	case "incremental":
+		return PgBackRestIncr, nil
+	default:
+		return "", mterrors.New(mtrpcpb.Code_INVALID_ARGUMENT,
+			fmt.Sprintf("invalid backup type '%s': must be one of: full, differential, incremental", backupType))
+	}
+}
