+++ go/services/multipooler/internal/manager/backup/check.go
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-// http://www.apache.org/licenses/LICENSE-2.0
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,49 +12,49 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package manager
+package backup
 
 import (
 	"context"
 	"fmt"
 
-	"github.com/multigres/multigres/go/common/backup"
+	commonbackup "github.com/multigres/multigres/go/common/backup"
 	"github.com/multigres/multigres/go/common/mterrors"
 	mtrpcpb "github.com/multigres/multigres/go/pb/mtrpc"
 	"github.com/multigres/multigres/go/tools/telemetry"
 )
 
-// runPgBackRestCheck runs `pgbackrest check`, which forces a WAL switch and
+// Check runs `pgbackrest check`, which forces a WAL switch and
 // confirms the segment reaches the repo, validating archive_command and
 // archive_mode end-to-end. On a standby pgBackRest skips the WAL switch, so
 // this is safe to call regardless of role. On failure the combined command
 // output is folded into the returned error.
-func (pm *MultiPoolerManager) runPgBackRestCheck(ctx context.Context) error {
-	configPath, err := pm.pgBackRestConfig()
+func (e *Engine) Check(ctx context.Context) error {
+	configPath, err := e.requireConfigPath()
 	if err != nil {
 		return mterrors.Wrap(err, "pgbackrest config not found")
 	}
 
-	ctx, cancel := context.WithTimeout(ctx, backup.CheckTimeout)
+	ctx, cancel := context.WithTimeout(ctx, commonbackup.CheckTimeout)
 	defer cancel()
 
 	args := []string{
-		"--stanza=" + pm.stanzaName(),
+		"--stanza=" + stanzaName,
 		"--config=" + configPath,
 		"check",
 	}
-	cmd := pm.pgbackrestCmd(ctx, args...)
+	cmd := e.pgbackrestCmd(ctx, args...)
 
 	var output []byte
 	err = telemetry.WithSpan(ctx, "pgbackrest-check", func(ctx context.Context) error {
 		var runErr error
-		output, runErr = pm.runLongCommand(ctx, cmd, "pgbackrest check")
+		output, runErr = e.run(ctx, cmd, "pgbackrest check")
 		return runErr
 	})
 	if err != nil {
 		return mterrors.New(mtrpcpb.Code_INTERNAL,
 			fmt.Sprintf("pgbackrest check failed for stanza %s: %v\nOutput: %s",
-				pm.stanzaName(), err, string(output)))
+				stanzaName, err, string(output)))
 	}
 	return nil
 }
