+++ go/services/multipooler/internal/manager/backup/types.go
@@ -1,0 +1,55 @@
+// Copyright 2026 Supabase, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package backup
+
+// pgBackRestInfo represents the structure of pgbackrest info JSON output
+type pgBackRestInfo struct {
+	Name   string             `json:"name"`
+	Backup []pgBackRestBackup `json:"backup"`
+}
+
+// pgBackRestBackup represents a single backup in the info output
+type pgBackRestBackup struct {
+	Label      string              `json:"label"`
+	Type       string              `json:"type"`
+	Error      bool                `json:"error"`
+	Timestamp  pgBackRestTimestamp `json:"timestamp"`
+	Annotation map[string]string   `json:"annotation,omitempty"`
+	LSN        *pgBackRestLSN      `json:"lsn,omitempty"` // LSN range for this backup
+	Info       *pgBackRestSizeInfo `json:"info,omitempty"`
+}
+
+// pgBackRestSizeInfo represents size information for a backup
+type pgBackRestSizeInfo struct {
+	Size       uint64                    `json:"size"`       // Original database size in bytes
+	Repository *pgBackRestRepositorySize `json:"repository"` // Compressed size in repository
+}
+
+// pgBackRestRepositorySize represents repository size info
+type pgBackRestRepositorySize struct {
+	Size uint64 `json:"size"` // Compressed size in repository in bytes
+}
+
+// pgBackRestTimestamp represents backup timestamps
+type pgBackRestTimestamp struct {
+	Start int64 `json:"start"`
+	Stop  int64 `json:"stop"`
+}
+
+// pgBackRestLSN represents LSN start/stop positions for a backup
+type pgBackRestLSN struct {
+	Start string `json:"start"` // e.g., "0/21000028"
+	Stop  string `json:"stop"`  // e.g., "0/21000100"
+}
