+++ go/services/multipooler/internal/manager/manager.go
@@ -38,11 +38,11 @@ import (
 	"github.com/multigres/multigres/go/services/multipooler/internal/executor"
 	"github.com/multigres/multigres/go/services/multipooler/internal/heartbeat"
 	"github.com/multigres/multigres/go/services/multipooler/internal/manager/actionlock"
+	backupengine "github.com/multigres/multigres/go/services/multipooler/internal/manager/backup"
 	"github.com/multigres/multigres/go/services/multipooler/internal/manager/consensus"
 	"github.com/multigres/multigres/go/services/multipooler/internal/poolerserver"
 	"github.com/multigres/multigres/go/services/multipooler/internal/pubsub"
 	"github.com/multigres/multigres/go/tools/ctxutil"
-	"github.com/multigres/multigres/go/tools/executil"
 	"github.com/multigres/multigres/go/tools/grpccommon"
 	"github.com/multigres/multigres/go/tools/retry"
 	"github.com/multigres/multigres/go/tools/telemetry"
@@ -136,10 +136,6 @@ type MultiPoolerManager struct {
 	// and broadcasts to all receivers) rather than sending to it.
 	readyChan chan struct{}
 
-	// Cached backup config from the database topology record.
-	// This is loaded once during startup and cached for fast access.
-	backupConfig *backup.Config
-
 	// pgBackRestConfigPath is the path to the pgbackrest.conf client config file
 	// generated by multipooler during startup. Set once in loadShardConfigFromGlobalTopo.
 	pgBackRestConfigPath string
@@ -205,8 +201,9 @@ type MultiPoolerManager struct {
 	// (query service, heartbeat tracker) and updates the multipooler record.
 	servingState *StateManager
 
-	// metrics holds OTel gauges for pgBackRest server health.
-	metrics *Metrics
+	// backup is the pgBackRest engine. The manager orchestrates lifecycle and
+	// owns the gRPC handlers, delegating the pgBackRest steps to the engine.
+	backup *backupengine.Engine
 
 	// healthStreamer streams health state to subscribers.
 	// Owns all health-related state and provides typed update methods.
@@ -341,11 +338,21 @@ func NewMultiPoolerManagerWithTimeout(logger *slog.Logger, multiPooler *clusterm
 	// The ReplTracker is registered later when heartbeat is started.
 	pm.servingState = NewStateManager(logger, pm.record, pm.qsc, pm.healthStreamer)
 
-	// Register pgBackRest health metrics.
-	var metricsErr error
-	pm.metrics, metricsErr = NewMetrics()
-	if metricsErr != nil {
-		logger.Warn("failed to register pgBackRest metrics", "error", metricsErr)
+	// Construct the pgBackRest engine. It registers the pgBackRest health
+	// metrics and owns all pgBackRest interaction. The pgbackrest.conf path,
+	// pgpass file, and repo config are supplied later (SetConfigPath /
+	// SetPgpassPath / SetBackupConfig) once topology and the backup location
+	// have loaded.
+	var engineErr error
+	pm.backup, engineErr = backupengine.NewEngine(pm.logger, pm.runLongCommand, pm.record, backupengine.Settings{
+		PgpassPath: pm.pgpassPath,
+		CAFile:     pm.config.PgBackRestCAFile,
+		CertFile:   pm.config.PgBackRestCertFile,
+		KeyFile:    pm.config.PgBackRestKeyFile,
+		PgDataDir:  postgresDataDir(),
+	})
+	if engineErr != nil {
+		logger.Warn("failed to construct pgBackRest engine", "error", engineErr)
 	}
 
 	return pm, nil
@@ -752,22 +759,6 @@ func (pm *MultiPoolerManager) GetStateAndError() (ManagerState, error) {
 	return pm.state, pm.stateError
 }
 
-// stanzaName returns the pgbackrest stanza name
-func (pm *MultiPoolerManager) stanzaName() string {
-	return "multigres"
-}
-
-// pgbackrestCmd builds a pgbackrest executil.Cmd with PGPASSFILE set so
-// pgbackrest can authenticate against the local PostgreSQL instance without
-// exposing the password in the process environment.
-func (pm *MultiPoolerManager) pgbackrestCmd(ctx context.Context, args ...string) *executil.Cmd {
-	cmd := executil.Command(ctx, "pgbackrest", args...)
-	if pm.pgpassPath != "" {
-		cmd.AddEnv("PGPASSFILE=" + pm.pgpassPath)
-	}
-	return cmd
-}
-
 // getPgCtldClient returns the pgctld gRPC client
 func (pm *MultiPoolerManager) getPgCtldClient() pgctldpb.PgCtldClient {
 	return pm.pgctldClient
@@ -1017,11 +1008,15 @@ func (pm *MultiPoolerManager) loadShardConfigFromGlobalTopo() {
 		}
 
 		pm.mu.Lock()
-		pm.backupConfig = backupConfig
 		pm.pgBackRestConfigPath = configPath
 		pm.pgpassPath = pgpassPath
 		pm.topoLoaded = true
 		pm.mu.Unlock()
+
+		// Feed the resolved config to the backup engine.
+		pm.backup.SetBackupConfig(backupConfig)
+		pm.backup.SetConfigPath(configPath)
+		pm.backup.SetPgpassPath(pgpassPath)
 
 		// Note: restoring from backup (for replicas) happens in a separate goroutine
 
@@ -2109,7 +2104,7 @@ func (pm *MultiPoolerManager) takeRemedialAction(ctx context.Context, action rem
 // hasCompleteBackups checks if there are any complete backups available
 func (pm *MultiPoolerManager) hasCompleteBackups(ctx context.Context) bool {
 	// Get list of backups
-	backups, err := pm.listBackups(ctx)
+	backups, err := pm.backup.ListBackups(ctx)
 	if err != nil {
 		return false
 	}
@@ -2188,7 +2183,7 @@ func (pm *MultiPoolerManager) restoreAndStartPostgres(ctx context.Context) error
 	}
 
 	// Get the latest complete backup
-	backups, err := pm.listBackups(ctx)
+	backups, err := pm.backup.ListBackups(ctx)
 	if err != nil {
 		return fmt.Errorf("failed to list backups: %w", err)
 	}
