+++ go/services/multipooler/internal/manager/manager_test.go
@@ -445,12 +445,12 @@ func TestGetBackupLocation(t *testing.T) {
 		utils.FilesystemBackupLocation("/var/backups/pgbackrest"),
 	)
 	require.NoError(t, err)
-	manager.backupConfig = backupConfig
+	manager.backup.SetBackupConfig(backupConfig)
 
 	// Test backup config
-	assert.Equal(t, "filesystem", manager.backupConfig.Type())
+	assert.Equal(t, "filesystem", backupConfig.Type())
 	expectedShardBackupLocation := filepath.Join("/var/backups/pgbackrest", database, constants.DefaultTableGroup, constants.DefaultShard)
-	shardPath, err := manager.backupConfig.FullPath(database, constants.DefaultTableGroup, constants.DefaultShard)
+	shardPath, err := backupConfig.FullPath(database, constants.DefaultTableGroup, constants.DefaultShard)
 	require.NoError(t, err)
 	assert.Equal(t, expectedShardBackupLocation, shardPath)
 }
@@ -497,19 +497,19 @@ func TestGetBackupLocation_S3(t *testing.T) {
 			utils.WithS3KeyPrefix("prod/backups/")),
 	)
 	require.NoError(t, err)
-	manager.backupConfig = backupConfig
+	manager.backup.SetBackupConfig(backupConfig)
 
 	// Test S3 backup config
-	assert.Equal(t, "s3", manager.backupConfig.Type())
+	assert.Equal(t, "s3", backupConfig.Type())
 
 	// Verify full path includes S3 bucket, prefix, and path components
 	expectedPath := "s3://my-backup-bucket/prod/backups/testdb/default/0-inf"
-	shardPath, err := manager.backupConfig.FullPath(database, constants.DefaultTableGroup, constants.DefaultShard)
+	shardPath, err := backupConfig.FullPath(database, constants.DefaultTableGroup, constants.DefaultShard)
 	require.NoError(t, err)
 	assert.Equal(t, expectedPath, shardPath)
 
 	// Verify PgBackRestConfig returns correct S3 settings
-	pgbrConfig, err := manager.backupConfig.PgBackRestConfig("multigres")
+	pgbrConfig, err := backupConfig.PgBackRestConfig("multigres")
 	require.NoError(t, err)
 	assert.Equal(t, "s3", pgbrConfig["repo1-type"])
 	assert.Equal(t, "my-backup-bucket", pgbrConfig["repo1-s3-bucket"])
