+++ go/services/multipooler/internal/manager/rpc_backup.go
@@ -16,38 +16,22 @@ package manager
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"log/slog"
-	"strings"
 	"time"
 
 	"github.com/multigres/multigres/go/common/backup"
 	"github.com/multigres/multigres/go/common/eventlog"
 	"github.com/multigres/multigres/go/common/mterrors"
-	"github.com/multigres/multigres/go/common/topoclient"
 	clustermetadatapb "github.com/multigres/multigres/go/pb/clustermetadata"
 	mtrpcpb "github.com/multigres/multigres/go/pb/mtrpc"
 	multipoolermanagerdata "github.com/multigres/multigres/go/pb/multipoolermanagerdata"
 	pgctldpb "github.com/multigres/multigres/go/pb/pgctldservice"
 	"github.com/multigres/multigres/go/services/multipooler/internal/manager/actionlock"
+	backupengine "github.com/multigres/multigres/go/services/multipooler/internal/manager/backup"
+	"github.com/multigres/multigres/go/tools/executil"
 	"github.com/multigres/multigres/go/tools/telemetry"
 )
-
-// pgBackRestConfig returns the path to the pgbackrest.conf generated by multipooler
-// during startup. The config is generated in loadShardConfigFromGlobalTopo when backup
-// location is loaded from topology.
-func (pm *MultiPoolerManager) pgBackRestConfig() (string, error) {
-	pm.mu.Lock()
-	configPath := pm.pgBackRestConfigPath
-	pm.mu.Unlock()
-
-	if configPath == "" {
-		return "", errors.New("pgbackrest config not yet generated (topology not loaded)")
-	}
-	return configPath, nil
-}
 
 // Backup performs a backup
 func (pm *MultiPoolerManager) Backup(ctx context.Context, forcePrimary bool, backupType string, jobID string, overrides map[string]string) (string, error) {
@@ -60,7 +44,7 @@ func (pm *MultiPoolerManager) Backup(ctx context.Context, forcePrimary bool, bac
 	var err error
 	lockStart := time.Now()
 	ctx, err = pm.actionLock.Acquire(ctx, "Backup")
-	pm.metrics.RecordBackupLockWait(ctx, time.Since(lockStart).Seconds())
+	pm.backup.Metrics().RecordBackupLockWait(ctx, time.Since(lockStart).Seconds())
 	if err != nil {
 		return "", err
 	}
@@ -83,558 +67,28 @@ func (pm *MultiPoolerManager) Backup(ctx context.Context, forcePrimary bool, bac
 
 // backupLocked performs a backup. Caller must hold the action lock and backup lease.
 func (pm *MultiPoolerManager) backupLocked(ctx context.Context, forcePrimary bool, backupType string, jobID string, overrides map[string]string) (retBackupID string, retErr error) {
-	retErr = telemetry.WithSpan(ctx, "backup", func(ctx context.Context) error {
-		var err error
-		retBackupID, err = pm.backupLockedInner(ctx, forcePrimary, backupType, jobID, overrides)
-		return err
-	})
-	return retBackupID, retErr
-}
-
-// backupLockedInner performs the backup steps inside the parent "backup" span.
-// Caller must hold both the action lock and the backup lease.
-func (pm *MultiPoolerManager) backupLockedInner(ctx context.Context, forcePrimary bool, backupType string, jobID string, overrides map[string]string) (retBackupID string, retErr error) {
-	if err := actionlock.AssertActionLockHeld(ctx); err != nil {
-		return "", err
-	}
-	if err := topoclient.AssertBackupLockHeld(ctx, pm.shardKey()); err != nil {
-		return "", mterrors.Wrap(err, "backup lease not held")
-	}
-
-	pm.metrics.IncBackupAttempts(ctx)
-	defer func() {
-		if retErr == nil {
-			pm.metrics.IncBackupSuccesses(ctx)
-		} else {
-			pm.metrics.IncBackupFailures(ctx)
-		}
-	}()
-
-	pm.logger.InfoContext(ctx, "Starting backup operation", "backup_type", pm.backupConfig.Type(), "job_type", backupType)
-
-	// Check if backup is allowed on primary
+	// Policy and primary-source resolution are owned by the manager (it has the
+	// role and the consensus-recorded primary); the engine is pure pgBackRest.
 	if err := pm.allowBackupOnPrimary(ctx, forcePrimary); err != nil {
 		return "", err
 	}
-
-	// Get path to pgbackrest.conf generated by pgctld
-	configPath, err := pm.pgBackRestConfig()
+	// Validate the requested backup type before resolving the backup source, so
+	// an invalid type is rejected up front rather than masked by a pg2 error.
+	pgBackRestType, err := backupengine.ValidateBackupType(backupType)
 	if err != nil {
 		return "", err
 	}
-
-	tableGroup := pm.record.ShardKey().GetTableGroup()
-	shard := pm.record.ShardKey().GetShard()
-	multipoolerID := topoclient.MultiPoolerIDString(pm.record.Id())
-	multipoolerName := pm.record.Id().Name
-
-	// Use provided job_id or generate one (same format as multiadmin)
-	effectiveJobID := jobID
-	if effectiveJobID == "" {
-		effectiveJobID = backup.GenerateJobID(multipoolerID)
-	}
-
-	// Validate parameters and get pgbackrest type
-	pgBackRestType, err := pm.validateBackupParams(backupType, configPath)
-	if err != nil {
-		return "", err
-	}
-
-	// Emit Started now that we have a correlation ID and parameters are valid.
-	// BackupName uses effectiveJobID since the actual pgbackrest label isn't assigned yet.
-	eventlog.Emit(ctx, pm.logger, eventlog.Started, eventlog.BackupAttempt{BackupName: effectiveJobID})
-	defer func() {
-		if retErr == nil {
-			// retBackupID is set to the actual pgbackrest label by the success return.
-			eventlog.Emit(ctx, pm.logger, eventlog.Success, eventlog.BackupAttempt{BackupName: retBackupID})
-		} else {
-			eventlog.Emit(ctx, pm.logger, eventlog.Failed, eventlog.BackupAttempt{BackupName: effectiveJobID}, "error", retErr)
-		}
-	}()
-
-	// Execute pgbackrest backup command
-	ctx, cancel := context.WithTimeout(ctx, backup.BackupTimeout)
-	defer cancel()
-
-	args := []string{
-		"--stanza=" + pm.stanzaName(),
-		"--config=" + configPath,
-		"--type=" + pgBackRestType,
-	}
-
-	// Get pg2 args for replica backups (empty for primary).
-	// For TLS mode, we need the PRIMARY's pgBackRest port, not the local one.
-	// Pass forcePrimary through so that a freshly-initialised node (type=UNKNOWN)
-	// that is acting as a primary can still do a local backup without pg2 args.
 	pg2Args, err := pm.GetPrimaryAsPg2Args(ctx, overrides, forcePrimary)
 	if err != nil {
 		return "", mterrors.Wrap(err, "failed to get primary as pg2 arguments")
 	}
-	args = append(args, pg2Args...)
 
-	// Reject backups with empty table_group or shard to prevent corrupt metadata.
-	// Backups missing these annotations would be silently skipped by replicas
-	// during restore, causing them to fall back to older backups.
-	if tableGroup == "" {
-		return "", mterrors.New(mtrpcpb.Code_FAILED_PRECONDITION, "table_group is missing")
-	}
-	if shard == "" {
-		return "", mterrors.New(mtrpcpb.Code_FAILED_PRECONDITION, "shard is missing")
-	}
-	args = append(args, "--annotation=table_group="+tableGroup)
-	args = append(args, "--annotation=shard="+shard)
-
-	// Add multipooler_id, pooler_type, and job_id annotations for unique identification
-	args = append(args, "--annotation=multipooler_id="+multipoolerName)
-	poolerType := pm.getPoolerType()
-	args = append(args, "--annotation=pooler_type="+poolerType.String())
-	args = append(args, "--annotation=job_id="+effectiveJobID)
-
-	args = append(args, "backup")
-
-	cmd := pm.pgbackrestCmd(ctx, args...)
-
-	// Execute backup with progress logging
-	var output []byte
-	backupStart := time.Now()
-	err = telemetry.WithSpan(ctx, "backup/pgbackrest", func(ctx context.Context) error {
-		var runErr error
-		output, runErr = pm.runLongCommand(ctx, cmd, "pgbackrest backup")
-		return runErr
+	retErr = telemetry.WithSpan(ctx, "backup", func(ctx context.Context) error {
+		var err error
+		retBackupID, err = pm.backup.Backup(ctx, pgBackRestType, jobID, pg2Args, pm.getPoolerType())
+		return err
 	})
-	pm.metrics.RecordBackupDuration(ctx, time.Since(backupStart).Seconds())
-	if err != nil {
-		return "", mterrors.New(mtrpcpb.Code_INTERNAL,
-			fmt.Sprintf("pgbackrest backup failed: %v\nOutput: %s", err, string(output)))
-	}
-
-	// Find the backup ID by querying pgbackrest info with our unique annotations
-	var foundBackupID string
-	err = telemetry.WithSpan(ctx, "backup/find", func(ctx context.Context) error {
-		var findErr error
-		foundBackupID, findErr = pm.findBackupByJobID(ctx, effectiveJobID)
-		return findErr
-	})
-	if err != nil {
-		return "", mterrors.New(mtrpcpb.Code_INTERNAL,
-			fmt.Sprintf("failed to find backup by annotations: %v\nBackup output: %s", err, string(output)))
-	}
-
-	// Verify the backup to ensure it's valid
-	verifyCtx, verifyCancel := context.WithTimeout(ctx, backup.VerifyTimeout)
-	defer verifyCancel()
-
-	verifyCmd := pm.pgbackrestCmd(verifyCtx,
-		"--stanza="+pm.stanzaName(),
-		"--config="+configPath,
-		"--set="+foundBackupID,
-		"verify")
-
-	// Execute verify with progress logging
-	var verifyOutput []byte
-	verifyStart := time.Now()
-	verifyErr := telemetry.WithSpan(verifyCtx, "backup/verify", func(ctx context.Context) error {
-		var runErr error
-		verifyOutput, runErr = pm.runLongCommand(ctx, verifyCmd, "pgbackrest verify backup_id="+foundBackupID)
-		return runErr
-	})
-	pm.metrics.RecordBackupVerifyDuration(ctx, time.Since(verifyStart).Seconds())
-	if verifyErr != nil {
-		return "", mterrors.New(mtrpcpb.Code_INTERNAL,
-			fmt.Sprintf("pgbackrest verify failed for backup %s: %v\nOutput: %s", foundBackupID, verifyErr, string(verifyOutput)))
-	}
-	// TODO: use `pgbackrest info` to verify that the database pages from the backed up
-	// Postgres cluster pass checksum validation.
-
-	return foundBackupID, nil
-}
-
-// RestoreFromBackup restores from a backup to a standby without a data directory.
-//
-// Requirements:
-// - The pooler must be a standby (not a primary)
-// - PGDATA must not exist (caller's responsibility to stop PostgreSQL and remove it)
-//
-// This function will:
-// 1. Execute pgbackrest restore to recreate PGDATA
-// 2. Start PostgreSQL in standby mode using Restart (which handles the not-running case)
-// 3. Reopen the pooler manager to establish fresh connections
-func (pm *MultiPoolerManager) RestoreFromBackup(ctx context.Context, backupID string) error {
-	slog.InfoContext(ctx, "RestoreFromBackup called", "backup_id", backupID)
-
-	// We can't proceed without the topo, which is loaded asynchronously at startup
-	if err := pm.checkReady(); err != nil {
-		return err
-	}
-
-	// Acquire the action lock to ensure only one mutation runs at a time
-	var err error
-	ctx, err = pm.actionLock.Acquire(ctx, "RestoreFromBackup")
-	if err != nil {
-		return err
-	}
-	defer pm.actionLock.Release(ctx)
-
-	return telemetry.WithSpan(ctx, "restore-from-backup", func(ctx context.Context) error {
-		return pm.restoreFromBackupLocked(ctx, backupID)
-	})
-}
-
-// restoreFromBackupLocked performs the restore. Caller must hold the action lock.
-func (pm *MultiPoolerManager) restoreFromBackupLocked(ctx context.Context, backupID string) (retErr error) {
-	if err := actionlock.AssertActionLockHeld(ctx); err != nil {
-		return err
-	}
-
-	pm.metrics.IncRestoreAttempts(ctx)
-	restoreStart := time.Now()
-	defer func() {
-		pm.metrics.RecordRestoreDuration(ctx, time.Since(restoreStart).Seconds())
-		if retErr == nil {
-			pm.metrics.IncRestoreSuccesses(ctx)
-		} else {
-			pm.metrics.IncRestoreFailures(ctx)
-		}
-	}()
-
-	pm.logger.InfoContext(ctx, "Starting restore operation", "backup_type", pm.backupConfig.Type(), "backup_id", backupID)
-
-	eventlog.Emit(ctx, pm.logger, eventlog.Started, eventlog.RestoreAttempt{BackupName: backupID})
-	defer func() {
-		if retErr == nil {
-			eventlog.Emit(ctx, pm.logger, eventlog.Success, eventlog.RestoreAttempt{BackupName: backupID})
-		} else {
-			eventlog.Emit(ctx, pm.logger, eventlog.Failed, eventlog.RestoreAttempt{BackupName: backupID}, "error", retErr)
-		}
-	}()
-
-	// Check that this is a standby, not a primary
-	poolerType := pm.getPoolerType()
-	if poolerType == clustermetadatapb.PoolerType_PRIMARY {
-		return mterrors.New(mtrpcpb.Code_FAILED_PRECONDITION,
-			"cannot restore to a primary pooler; restore is only supported for standby poolers")
-	}
-
-	// Check that PGDATA doesn't exist (caller must remove it before restore)
-	if pm.hasDataDirectory() {
-		return mterrors.New(mtrpcpb.Code_FAILED_PRECONDITION,
-			"cannot restore: PGDATA already exists; caller must stop PostgreSQL and remove PGDATA first")
-	}
-
-	// Restore the backup
-	if err := telemetry.WithSpan(ctx, "restore/pgbackrest", func(ctx context.Context) error {
-		return pm.executePgBackrestRestore(ctx, backupID)
-	}); err != nil {
-		return err
-	}
-
-	// Reconfigure archive_command to use this pooler's local pgbackrest.conf
-	// After restore, postgresql.auto.conf contains archive_command pointing to the
-	// primary's pgbackrest.conf path. Each pooler needs its own config path.
-	pm.logger.InfoContext(ctx, "Reconfiguring archive_command for local pgbackrest.conf")
-	if err := telemetry.WithSpan(ctx, "restore/reconfigure-archive", func(ctx context.Context) error {
-		if err := pm.removeArchiveConfigFromAutoConf(); err != nil {
-			return mterrors.Wrap(err, "failed to remove old archive configuration")
-		}
-		return mterrors.Wrap(pm.configureArchiveMode(ctx), "failed to configure archive mode")
-	}); err != nil {
-		return err
-	}
-
-	if err := telemetry.WithSpan(ctx, "restore/start-postgres", func(ctx context.Context) error {
-		return pm.startPostgreSQLAfterRestore(ctx, backupID)
-	}); err != nil {
-		return err
-	}
-
-	// Clear the in-memory leader observation. The restored PGDATA may have been
-	// from a different point in time, so the previously observed leader may no
-	// longer be accurate. The term revocation is intentionally preserved: it is
-	// monotonically increasing and the restore does not change who is allowed to
-	// lead — only a coordinator with a term >= the revocation term may configure
-	// this node, which is exactly the right safety property.
-	pm.healthStreamer.UpdateLeaderObservation(nil)
-
-	if err := telemetry.WithSpan(ctx, "restore/reopen-pooler", func(ctx context.Context) error {
-		return pm.reopenPoolerManager(ctx)
-	}); err != nil {
-		return err
-	}
-
-	// Mark as initialized after successful restore
-	return telemetry.WithSpan(ctx, "restore/mark-initialized", func(_ context.Context) error {
-		if err := pm.setInitialized(); err != nil {
-			return err
-		}
-		// Push an immediate health snapshot so the orchestrator learns about
-		// IsInitialized=true without waiting for the next 30-second heartbeat.
-		pm.broadcastHealth()
-		return nil
-	})
-}
-
-func (pm *MultiPoolerManager) executePgBackrestRestore(ctx context.Context, backupID string) error {
-	configPath, err := pm.pgBackRestConfig()
-	if err != nil {
-		return mterrors.Wrap(err, "pgbackrest config not found")
-	}
-
-	restoreCtx, cancel := context.WithTimeout(ctx, backup.RestoreTimeout)
-	defer cancel()
-
-	args := []string{
-		"--stanza=" + pm.stanzaName(),
-		"--config=" + configPath,
-		"--type=standby",
-	}
-
-	if backupID != "" {
-		args = append(args, "--set="+backupID)
-	}
-
-	args = append(args, "restore")
-
-	cmd := pm.pgbackrestCmd(restoreCtx, args...)
-	output, err := safeCombinedOutput(cmd)
-	if err != nil {
-		return mterrors.New(mtrpcpb.Code_INTERNAL,
-			fmt.Sprintf("pgbackrest restore failed: %v\nOutput: %s", err, output))
-	}
-
-	return nil
-}
-
-func (pm *MultiPoolerManager) startPostgreSQLAfterRestore(ctx context.Context, backupID string) error {
-	pgctldClient := pm.getPgCtldClient()
-	if pgctldClient == nil {
-		return mterrors.New(mtrpcpb.Code_INVALID_ARGUMENT, "pgctld_client is required")
-	}
-
-	restartCtx, cancel := context.WithTimeout(ctx, 2*time.Minute)
-	defer cancel()
-
-	slog.InfoContext(ctx, "Starting PostgreSQL after restore",
-		"backup_id", backupID)
-
-	_, err := pgctldClient.Restart(restartCtx, &pgctldpb.RestartRequest{
-		AsStandby: true,
-	})
-	if err != nil {
-		return mterrors.New(mtrpcpb.Code_INTERNAL,
-			fmt.Sprintf("failed to start PostgreSQL after restore: %v", err))
-	}
-
-	slog.InfoContext(ctx, "PostgreSQL started successfully after restore")
-	return nil
-}
-
-func (pm *MultiPoolerManager) reopenPoolerManager(ctx context.Context) error {
-	slog.InfoContext(ctx, "Reopening pooler manager after restore")
-	// Use reopenConnections instead of Pause/Open to avoid canceling pm.ctx.
-	// This is important during auto-restore at startup where the startup flow
-	// is waiting on contexts derived from pm.ctx.
-	pm.reopenConnections(ctx)
-	slog.InfoContext(ctx, "Pooler manager reopened successfully after restore")
-	return nil
-}
-
-// GetBackups retrieves backup information
-func (pm *MultiPoolerManager) GetBackups(ctx context.Context, limit uint32) ([]*multipoolermanagerdata.BackupMetadata, error) {
-	// We can't proceed without the topo, which is loaded asynchronously at startup
-	if err := pm.checkReady(); err != nil {
-		return nil, err
-	}
-
-	// Acquire the action lock to ensure only one operation runs at a time
-	var err error
-	ctx, err = pm.actionLock.Acquire(ctx, "GetBackups")
-	if err != nil {
-		return nil, err
-	}
-	defer pm.actionLock.Release(ctx)
-
-	return pm.getBackupsLocked(ctx, limit)
-}
-
-// getBackupsLocked retrieves backup information. Caller must hold the action lock.
-func (pm *MultiPoolerManager) getBackupsLocked(ctx context.Context, limit uint32) ([]*multipoolermanagerdata.BackupMetadata, error) {
-	if err := actionlock.AssertActionLockHeld(ctx); err != nil {
-		return nil, err
-	}
-
-	backups, err := pm.listBackups(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	// Apply limit if specified
-	if limit > 0 && uint32(len(backups)) > limit {
-		backups = backups[:limit]
-	}
-
-	return backups, nil
-}
-
-// listBackups retrieves backup metadata from pgbackrest.
-// Caller must hold the action lock.
-// pm.mu is NOT required because this function only reads config fields
-// (which are immutable after construction) and backupLocation (which is
-// set once during startup and never modified).
-func (pm *MultiPoolerManager) listBackups(ctx context.Context) ([]*multipoolermanagerdata.BackupMetadata, error) {
-	configPath, err := pm.pgBackRestConfig()
-	if err != nil {
-		return nil, mterrors.Wrap(err, "pgbackrest config not found")
-	}
-
-	if pm.backupConfig == nil {
-		return nil, mterrors.New(mtrpcpb.Code_INVALID_ARGUMENT, "backup config not loaded from topology")
-	}
-
-	// Execute pgbackrest info command with JSON output
-	queryCtx, cancel := context.WithTimeout(ctx, backup.InfoTimeout)
-	defer cancel()
-
-	cmd := pm.pgbackrestCmd(queryCtx,
-		"--stanza="+pm.stanzaName(),
-		"--config="+configPath,
-		"--output=json",
-		"--log-level-console=off", // Override console logging to prevent contaminating JSON output
-		"info")
-
-	output, err := safeCombinedOutput(cmd)
-	if err != nil {
-		// Handle case where stanza doesn't exist yet or config file is missing - return empty list
-		if output == "" || strings.Contains(output, "does not exist") || strings.Contains(output, "unable to open missing file") {
-			return []*multipoolermanagerdata.BackupMetadata{}, nil
-		}
-		return nil, mterrors.New(mtrpcpb.Code_INTERNAL,
-			fmt.Sprintf("pgbackrest info failed: %v\nOutput: %s", err, output))
-	}
-
-	// Parse JSON output
-	var infoData []pgBackRestInfo
-	if err := json.Unmarshal([]byte(output), &infoData); err != nil {
-		return nil, mterrors.New(mtrpcpb.Code_INTERNAL,
-			fmt.Sprintf("failed to parse pgbackrest info JSON: %v", err))
-	}
-
-	if len(infoData) == 0 || len(infoData[0].Backup) == 0 {
-		return []*multipoolermanagerdata.BackupMetadata{}, nil
-	}
-
-	// Get current pooler's table_group and shard for filtering
-	currentTableGroup := pm.record.ShardKey().GetTableGroup()
-	currentShard := pm.getShardID()
-
-	// Extract backups from the first stanza (should be the only one)
-	var backups []*multipoolermanagerdata.BackupMetadata
-	for _, pgBackup := range infoData[0].Backup {
-		status := multipoolermanagerdata.BackupMetadata_COMPLETE
-		if pgBackup.Error {
-			status = multipoolermanagerdata.BackupMetadata_INCOMPLETE
-		}
-
-		// Extract table_group, shard, job_id, multipooler_id, and pooler_type from annotations
-		tableGroup := ""
-		shard := ""
-		jobID := ""
-		multipoolerID := ""
-		poolerType := clustermetadatapb.PoolerType_UNKNOWN
-		if pgBackup.Annotation != nil {
-			tableGroup = pgBackup.Annotation["table_group"]
-			shard = pgBackup.Annotation["shard"]
-			jobID = pgBackup.Annotation["job_id"]
-			multipoolerID = pgBackup.Annotation["multipooler_id"]
-			if pt, ok := clustermetadatapb.PoolerType_value[pgBackup.Annotation["pooler_type"]]; ok {
-				poolerType = clustermetadatapb.PoolerType(pt)
-			}
-		}
-
-		// Defense-in-depth: skip backups that don't match this pooler's shard.
-		// This check is not strictly necessary since stanzas are shard-scoped,
-		// but provides an extra layer of safety.
-		if tableGroup != currentTableGroup {
-			pm.logger.ErrorContext(ctx, "Skipping backup with mismatched table_group",
-				"backup_id", pgBackup.Label,
-				"backup_table_group", tableGroup,
-				"current_table_group", currentTableGroup)
-			continue
-		}
-		if shard != currentShard {
-			pm.logger.ErrorContext(ctx, "Skipping backup with mismatched shard",
-				"backup_id", pgBackup.Label,
-				"backup_shard", shard,
-				"current_shard", currentShard)
-			continue
-		}
-
-		// Extract final LSN (stop LSN) from backup
-		finalLSN := ""
-		if pgBackup.LSN != nil && pgBackup.LSN.Stop != "" {
-			finalLSN = pgBackup.LSN.Stop
-		}
-
-		// Extract backup size if available
-		var backupSizeBytes uint64
-		if pgBackup.Info != nil {
-			backupSizeBytes = pgBackup.Info.Size
-		}
-
-		backups = append(backups, &multipoolermanagerdata.BackupMetadata{
-			BackupId:        pgBackup.Label,
-			Status:          status,
-			TableGroup:      tableGroup,
-			Shard:           shard,
-			FinalLsn:        finalLSN,
-			JobId:           jobID,
-			BackupSizeBytes: backupSizeBytes,
-			Type:            pgBackup.Type,
-			MultipoolerId:   multipoolerID,
-			PoolerType:      poolerType,
-		})
-	}
-
-	return backups, nil
-}
-
-// pgBackRestInfo represents the structure of pgbackrest info JSON output
-type pgBackRestInfo struct {
-	Name   string             `json:"name"`
-	Backup []pgBackRestBackup `json:"backup"`
-}
-
-// pgBackRestBackup represents a single backup in the info output
-type pgBackRestBackup struct {
-	Label      string              `json:"label"`
-	Type       string              `json:"type"`
-	Error      bool                `json:"error"`
-	Timestamp  pgBackRestTimestamp `json:"timestamp"`
-	Annotation map[string]string   `json:"annotation,omitempty"`
-	LSN        *pgBackRestLSN      `json:"lsn,omitempty"` // LSN range for this backup
-	Info       *pgBackRestSizeInfo `json:"info,omitempty"`
-}
-
-// pgBackRestSizeInfo represents size information for a backup
-type pgBackRestSizeInfo struct {
-	Size       uint64                    `json:"size"`       // Original database size in bytes
-	Repository *pgBackRestRepositorySize `json:"repository"` // Compressed size in repository
-}
-
-// pgBackRestRepositorySize represents repository size info
-type pgBackRestRepositorySize struct {
-	Size uint64 `json:"size"` // Compressed size in repository in bytes
-}
-
-// pgBackRestTimestamp represents backup timestamps
-type pgBackRestTimestamp struct {
-	Start int64 `json:"start"`
-	Stop  int64 `json:"stop"`
-}
-
-// pgBackRestLSN represents LSN start/stop positions for a backup
-type pgBackRestLSN struct {
-	Start string `json:"start"` // e.g., "0/21000028"
-	Stop  string `json:"stop"`  // e.g., "0/21000100"
+	return retBackupID, retErr
 }
 
 // allowBackupOnPrimary checks if a backup operation is allowed on a primary pooler
@@ -648,143 +102,6 @@ func (pm *MultiPoolerManager) allowBackupOnPrimary(ctx context.Context, forcePri
 			"backups from primary databases are not allowed unless ForcePrimary is set")
 	}
 	return nil
-}
-
-// validateBackupParams validates the backup type and returns the pgbackrest type mapping
-func (pm *MultiPoolerManager) validateBackupParams(backupType, configPath string) (pgBackRestType string, err error) {
-	// Validate backup type is provided
-	if backupType == "" {
-		return "", mterrors.New(mtrpcpb.Code_INVALID_ARGUMENT, "type is required")
-	}
-
-	// Map to pgbackrest types: full, diff, incr
-	switch backupType {
-	case "full":
-		pgBackRestType = "full"
-	case "differential":
-		pgBackRestType = "diff"
-	case "incremental":
-		pgBackRestType = "incr"
-	default:
-		return "", mterrors.New(mtrpcpb.Code_INVALID_ARGUMENT,
-			fmt.Sprintf("invalid backup type '%s': must be one of: full, differential, incremental", backupType))
-	}
-
-	// Validate required backup configuration
-	if configPath == "" {
-		return "", mterrors.New(mtrpcpb.Code_INVALID_ARGUMENT, "config_path is required")
-	}
-
-	return pgBackRestType, nil
-}
-
-// findBackupByJobID finds a backup by matching the job_id annotation
-//
-// This function has to scan the entire backup history to find the backup, but the worst case
-// should be manageable, because production deployments should have retention policies that
-// trigger pgBackRest to delete older backups.
-func (pm *MultiPoolerManager) findBackupByJobID(
-	ctx context.Context,
-	jobID string,
-) (string, error) {
-	configPath, err := pm.pgBackRestConfig()
-	if err != nil {
-		return "", mterrors.Wrap(err, "pgbackrest config not found")
-	}
-
-	// Execute pgbackrest info command with JSON output
-	infoCtx, cancel := context.WithTimeout(ctx, backup.InfoTimeout)
-	defer cancel()
-
-	cmd := pm.pgbackrestCmd(infoCtx,
-		"--stanza="+pm.stanzaName(),
-		"--config="+configPath,
-		"--output=json",
-		"--log-level-console=off",
-		"info")
-
-	output, err := safeCombinedOutput(cmd)
-	if err != nil {
-		return "", mterrors.New(mtrpcpb.Code_INTERNAL,
-			fmt.Sprintf("pgbackrest info failed: %v\nOutput: %s", err, output))
-	}
-
-	// Parse JSON output
-	var infoData []pgBackRestInfo
-	if err := json.Unmarshal([]byte(output), &infoData); err != nil {
-		return "", mterrors.New(mtrpcpb.Code_INTERNAL,
-			fmt.Sprintf("failed to parse pgbackrest info JSON: %v", err))
-	}
-
-	// Search for backup with matching annotations
-	if len(infoData) == 0 || len(infoData[0].Backup) == 0 {
-		return "", mterrors.New(mtrpcpb.Code_NOT_FOUND,
-			"no backups found in pgbackrest info output")
-	}
-
-	var matchedBackups []string
-	for _, pgBackup := range infoData[0].Backup {
-		if pgBackup.Annotation != nil {
-			if pgBackup.Annotation["job_id"] == jobID {
-				matchedBackups = append(matchedBackups, pgBackup.Label)
-			}
-		}
-	}
-
-	if len(matchedBackups) == 0 {
-		return "", mterrors.New(mtrpcpb.Code_NOT_FOUND,
-			"no backup found with job_id="+jobID)
-	}
-
-	if len(matchedBackups) > 1 {
-		return "", mterrors.New(mtrpcpb.Code_INTERNAL,
-			fmt.Sprintf("found %d backups with job_id=%s, expected 1",
-				len(matchedBackups), jobID))
-	}
-
-	return matchedBackups[0], nil
-}
-
-// GetBackupByJobId searches for a backup with the given job_id annotation.
-// Returns nil Backup if not found.
-func (pm *MultiPoolerManager) GetBackupByJobId(ctx context.Context, jobID string) (*multipoolermanagerdata.BackupMetadata, error) {
-	if jobID == "" {
-		return nil, mterrors.New(mtrpcpb.Code_INVALID_ARGUMENT, "job_id is required")
-	}
-
-	pm.logger.DebugContext(ctx, "Searching for backup by job_id", "job_id", jobID)
-
-	// We can't proceed without the topo, which is loaded asynchronously at startup
-	if err := pm.checkReady(); err != nil {
-		return nil, err
-	}
-
-	// Acquire the action lock to ensure only one operation runs at a time
-	var err error
-	ctx, err = pm.actionLock.Acquire(ctx, "GetBackupByJobId")
-	if err != nil {
-		return nil, err
-	}
-	defer pm.actionLock.Release(ctx)
-
-	// Get all backups and search for matching job_id
-	backups, err := pm.getBackupsLocked(ctx, 0) // 0 = no limit
-	if err != nil {
-		return nil, mterrors.Wrap(err, "failed to get backups")
-	}
-
-	for _, backup := range backups {
-		if backup.JobId == jobID {
-			pm.logger.DebugContext(ctx, "Found backup by job_id",
-				"job_id", jobID,
-				"backup_id", backup.BackupId,
-				"status", backup.Status)
-			return backup, nil
-		}
-	}
-
-	pm.logger.DebugContext(ctx, "Backup not found by job_id", "job_id", jobID)
-	return nil, nil
 }
 
 // GetPrimaryAsPg2Args returns pgbackrest CLI arguments for pg2 (primary) configuration.
@@ -887,4 +204,275 @@ func (pm *MultiPoolerManager) GetPrimaryAsPg2Args(
 	args = backup.ApplyPgBackRestOverrides(args, overrides)
 
 	return args, nil
+}
+
+// RestoreFromBackup restores from a backup to a standby without a data directory.
+//
+// Requirements:
+// - The pooler must be a standby (not a primary)
+// - PGDATA must not exist (caller's responsibility to stop PostgreSQL and remove it)
+//
+// This function will:
+// 1. Execute pgbackrest restore to recreate PGDATA
+// 2. Start PostgreSQL in standby mode using Restart (which handles the not-running case)
+// 3. Reopen the pooler manager to establish fresh connections
+func (pm *MultiPoolerManager) RestoreFromBackup(ctx context.Context, backupID string) error {
+	slog.InfoContext(ctx, "RestoreFromBackup called", "backup_id", backupID)
+
+	// We can't proceed without the topo, which is loaded asynchronously at startup
+	if err := pm.checkReady(); err != nil {
+		return err
+	}
+
+	// Acquire the action lock to ensure only one mutation runs at a time
+	var err error
+	ctx, err = pm.actionLock.Acquire(ctx, "RestoreFromBackup")
+	if err != nil {
+		return err
+	}
+	defer pm.actionLock.Release(ctx)
+
+	return telemetry.WithSpan(ctx, "restore-from-backup", func(ctx context.Context) error {
+		return pm.restoreFromBackupLocked(ctx, backupID)
+	})
+}
+
+// restoreFromBackupLocked performs the restore. Caller must hold the action lock.
+func (pm *MultiPoolerManager) restoreFromBackupLocked(ctx context.Context, backupID string) (retErr error) {
+	if err := actionlock.AssertActionLockHeld(ctx); err != nil {
+		return err
+	}
+
+	metrics := pm.backup.Metrics()
+	metrics.IncRestoreAttempts(ctx)
+	restoreStart := time.Now()
+	defer func() {
+		metrics.RecordRestoreDuration(ctx, time.Since(restoreStart).Seconds())
+		if retErr == nil {
+			metrics.IncRestoreSuccesses(ctx)
+		} else {
+			metrics.IncRestoreFailures(ctx)
+		}
+	}()
+
+	pm.logger.InfoContext(ctx, "Starting restore operation", "backup_id", backupID)
+
+	eventlog.Emit(ctx, pm.logger, eventlog.Started, eventlog.RestoreAttempt{BackupName: backupID})
+	defer func() {
+		if retErr == nil {
+			eventlog.Emit(ctx, pm.logger, eventlog.Success, eventlog.RestoreAttempt{BackupName: backupID})
+		} else {
+			eventlog.Emit(ctx, pm.logger, eventlog.Failed, eventlog.RestoreAttempt{BackupName: backupID}, "error", retErr)
+		}
+	}()
+
+	// Check that this is a standby, not a primary
+	poolerType := pm.getPoolerType()
+	if poolerType == clustermetadatapb.PoolerType_PRIMARY {
+		return mterrors.New(mtrpcpb.Code_FAILED_PRECONDITION,
+			"cannot restore to a primary pooler; restore is only supported for standby poolers")
+	}
+
+	// Check that PGDATA doesn't exist (caller must remove it before restore)
+	if pm.hasDataDirectory() {
+		return mterrors.New(mtrpcpb.Code_FAILED_PRECONDITION,
+			"cannot restore: PGDATA already exists; caller must stop PostgreSQL and remove PGDATA first")
+	}
+
+	// Restore the backup
+	if err := telemetry.WithSpan(ctx, "restore/pgbackrest", func(ctx context.Context) error {
+		return pm.backup.Restore(ctx, backupID)
+	}); err != nil {
+		return err
+	}
+
+	// Reconfigure archive_command to use this pooler's local pgbackrest.conf
+	// After restore, postgresql.auto.conf contains archive_command pointing to the
+	// primary's pgbackrest.conf path. Each pooler needs its own config path.
+	pm.logger.InfoContext(ctx, "Reconfiguring archive_command for local pgbackrest.conf")
+	if err := telemetry.WithSpan(ctx, "restore/reconfigure-archive", func(ctx context.Context) error {
+		if err := pm.backup.RemoveArchiveConfig(); err != nil {
+			return mterrors.Wrap(err, "failed to remove old archive configuration")
+		}
+		return mterrors.Wrap(pm.backup.ConfigureArchiveMode(ctx), "failed to configure archive mode")
+	}); err != nil {
+		return err
+	}
+
+	if err := telemetry.WithSpan(ctx, "restore/start-postgres", func(ctx context.Context) error {
+		return pm.startPostgreSQLAfterRestore(ctx, backupID)
+	}); err != nil {
+		return err
+	}
+
+	// Clear the in-memory leader observation. The restored PGDATA may have been
+	// from a different point in time, so the previously observed leader may no
+	// longer be accurate. The term revocation is intentionally preserved: it is
+	// monotonically increasing and the restore does not change who is allowed to
+	// lead — only a coordinator with a term >= the revocation term may configure
+	// this node, which is exactly the right safety property.
+	pm.healthStreamer.UpdateLeaderObservation(nil)
+
+	if err := telemetry.WithSpan(ctx, "restore/reopen-pooler", func(ctx context.Context) error {
+		return pm.reopenPoolerManager(ctx)
+	}); err != nil {
+		return err
+	}
+
+	// Mark as initialized after successful restore
+	return telemetry.WithSpan(ctx, "restore/mark-initialized", func(_ context.Context) error {
+		if err := pm.setInitialized(); err != nil {
+			return err
+		}
+		// Push an immediate health snapshot so the orchestrator learns about
+		// IsInitialized=true without waiting for the next 30-second heartbeat.
+		pm.broadcastHealth()
+		return nil
+	})
+}
+
+func (pm *MultiPoolerManager) startPostgreSQLAfterRestore(ctx context.Context, backupID string) error {
+	pgctldClient := pm.getPgCtldClient()
+	if pgctldClient == nil {
+		return mterrors.New(mtrpcpb.Code_INVALID_ARGUMENT, "pgctld_client is required")
+	}
+
+	restartCtx, cancel := context.WithTimeout(ctx, 2*time.Minute)
+	defer cancel()
+
+	slog.InfoContext(ctx, "Starting PostgreSQL after restore",
+		"backup_id", backupID)
+
+	_, err := pgctldClient.Restart(restartCtx, &pgctldpb.RestartRequest{
+		AsStandby: true,
+	})
+	if err != nil {
+		return mterrors.New(mtrpcpb.Code_INTERNAL,
+			fmt.Sprintf("failed to start PostgreSQL after restore: %v", err))
+	}
+
+	slog.InfoContext(ctx, "PostgreSQL started successfully after restore")
+	return nil
+}
+
+func (pm *MultiPoolerManager) reopenPoolerManager(ctx context.Context) error {
+	slog.InfoContext(ctx, "Reopening pooler manager after restore")
+	// Use reopenConnections instead of Pause/Open to avoid canceling pm.ctx.
+	// This is important during auto-restore at startup where the startup flow
+	// is waiting on contexts derived from pm.ctx.
+	pm.reopenConnections(ctx)
+	slog.InfoContext(ctx, "Pooler manager reopened successfully after restore")
+	return nil
+}
+
+// GetBackups retrieves backup information
+func (pm *MultiPoolerManager) GetBackups(ctx context.Context, limit uint32) ([]*multipoolermanagerdata.BackupMetadata, error) {
+	// We can't proceed without the topo, which is loaded asynchronously at startup
+	if err := pm.checkReady(); err != nil {
+		return nil, err
+	}
+
+	// Acquire the action lock to ensure only one operation runs at a time
+	var err error
+	ctx, err = pm.actionLock.Acquire(ctx, "GetBackups")
+	if err != nil {
+		return nil, err
+	}
+	defer pm.actionLock.Release(ctx)
+
+	if err := actionlock.AssertActionLockHeld(ctx); err != nil {
+		return nil, err
+	}
+	return pm.backup.List(ctx, limit)
+}
+
+// GetBackupByJobId searches for a backup with the given job_id annotation.
+// Returns nil Backup if not found.
+func (pm *MultiPoolerManager) GetBackupByJobId(ctx context.Context, jobID string) (*multipoolermanagerdata.BackupMetadata, error) {
+	if jobID == "" {
+		return nil, mterrors.New(mtrpcpb.Code_INVALID_ARGUMENT, "job_id is required")
+	}
+
+	pm.logger.DebugContext(ctx, "Searching for backup by job_id", "job_id", jobID)
+
+	// We can't proceed without the topo, which is loaded asynchronously at startup
+	if err := pm.checkReady(); err != nil {
+		return nil, err
+	}
+
+	// Acquire the action lock to ensure only one operation runs at a time
+	var err error
+	ctx, err = pm.actionLock.Acquire(ctx, "GetBackupByJobId")
+	if err != nil {
+		return nil, err
+	}
+	defer pm.actionLock.Release(ctx)
+
+	// Get all backups and search for matching job_id
+	backups, err := pm.backup.List(ctx, 0) // 0 = no limit
+	if err != nil {
+		return nil, mterrors.Wrap(err, "failed to get backups")
+	}
+
+	for _, backup := range backups {
+		if backup.JobId == jobID {
+			pm.logger.DebugContext(ctx, "Found backup by job_id",
+				"job_id", jobID,
+				"backup_id", backup.BackupId,
+				"status", backup.Status)
+			return backup, nil
+		}
+	}
+
+	pm.logger.DebugContext(ctx, "Backup not found by job_id", "job_id", jobID)
+	return nil, nil
+}
+
+// runLongCommand executes a long-running command with periodic progress logging.
+// Logs progress every 10 seconds. The cmd should be created with exec.CommandContext(ctx, ...)
+// to ensure proper cleanup on context cancellation.
+func (pm *MultiPoolerManager) runLongCommand(ctx context.Context, cmd *executil.Cmd, operationName string) ([]byte, error) {
+	pm.logger.InfoContext(ctx, "Starting command", "operation", operationName)
+
+	startTime := time.Now()
+
+	// Create a context for the logging goroutine
+	logCtx, cancelLog := context.WithCancel(ctx)
+
+	// Log progress periodically in background
+	go func() {
+		ticker := time.NewTicker(10 * time.Second)
+		defer ticker.Stop()
+
+		for {
+			select {
+			case <-logCtx.Done():
+				return
+			case <-ticker.C:
+				elapsed := time.Since(startTime)
+				pm.logger.InfoContext(ctx, "Command still in progress",
+					"operation", operationName,
+					"elapsed_seconds", int(elapsed.Seconds()))
+			}
+		}
+	}()
+
+	output, err := cmd.CombinedOutput()
+
+	cancelLog()
+
+	// Log completion
+	elapsed := time.Since(startTime)
+	if err != nil {
+		pm.logger.ErrorContext(ctx, "Command failed",
+			"operation", operationName,
+			"elapsed_seconds", int(elapsed.Seconds()),
+			"error", err)
+	} else {
+		pm.logger.InfoContext(ctx, "Command completed",
+			"operation", operationName,
+			"elapsed_seconds", int(elapsed.Seconds()))
+	}
+
+	return output, err
 }
