+++ go/services/multipooler/internal/manager/rpc_backup_test.go
@@ -15,12 +15,9 @@
 package manager
 
 import (
-	"bufio"
 	"context"
-	"errors"
 	"log/slog"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"testing"
@@ -34,9 +31,9 @@ import (
 	"github.com/multigres/multigres/go/common/topoclient"
 	"github.com/multigres/multigres/go/common/topoclient/memorytopo"
 	"github.com/multigres/multigres/go/services/multipooler/internal/manager/actionlock"
+	backupengine "github.com/multigres/multigres/go/services/multipooler/internal/manager/backup"
 	"github.com/multigres/multigres/go/services/multipooler/internal/manager/consensus"
 	"github.com/multigres/multigres/go/test/utils"
-	"github.com/multigres/multigres/go/tools/executil"
 	"github.com/multigres/multigres/go/tools/timer"
 
 	clustermetadatapb "github.com/multigres/multigres/go/pb/clustermetadata"
@@ -101,19 +98,25 @@ func createTestManagerWithBackupLocation(poolerDir, tableGroup, shard string, po
 	monitorRunner := timer.NewPeriodicRunner(context.TODO(), 10*time.Second)
 
 	pm := &MultiPoolerManager{
-		config:       &Config{},
-		serviceID:    multipoolerID,
-		topoClient:   topoClient,
-		record:       newRecordFromProto(multipoolerProto),
-		state:        ManagerStateReady,
-		backupConfig: backupConfig,
-		actionLock:   actionlock.NewActionLock(),
-		logger:       slog.Default(),
-		pgMonitor:    monitorRunner,
+		config:     &Config{},
+		serviceID:  multipoolerID,
+		topoClient: topoClient,
+		record:     newRecordFromProto(multipoolerProto),
+		state:      ManagerStateReady,
+		actionLock: actionlock.NewActionLock(),
+		logger:     slog.Default(),
+		pgMonitor:  monitorRunner,
 		// consensus.ConsensusState is the canonical home for the recorded primary
 		// (replacing the former pm.primaryHost/Port/PoolerID fields). Backup
 		// tests seed it via setBackupPrimary below.
 		consensusState: consensus.NewConsensusState(poolerDir, multipoolerID),
+	}
+
+	// Build the backup engine the way the production constructor does, feeding
+	// it the resolved repo config when one was supplied.
+	pm.backup, _ = backupengine.NewEngine(pm.logger, pm.runLongCommand, pm.record, backupengine.Settings{})
+	if backupConfig != nil {
+		pm.backup.SetBackupConfig(backupConfig)
 	}
 	return pm
 }
@@ -163,153 +166,6 @@ pg1-port=5432
 	err = os.WriteFile(configPath, []byte(configContent), 0o600)
 	require.NoError(t, err, "failed to create pgbackrest.conf")
 	return configPath
-}
-
-func TestFindBackupByJobID(t *testing.T) {
-	tests := []struct {
-		name          string
-		jobID         string
-		jsonOutput    string
-		wantBackupID  string
-		wantError     bool
-		errorContains string
-	}{
-		{
-			name:  "Single matching backup",
-			jobID: "20250104-100000.000000_mp-zone1",
-			jsonOutput: `[{
-				"backup": [{
-					"label": "20250104-100000F",
-					"annotation": {
-						"multipooler_id": "zone1-multipooler1",
-						"job_id": "20250104-100000.000000_mp-zone1"
-					}
-				}]
-			}]`,
-			wantBackupID: "20250104-100000F",
-			wantError:    false,
-		},
-		{
-			name:  "Multiple backups, one match",
-			jobID: "20250104-120000.000000_mp-zone1",
-			jsonOutput: `[{
-				"backup": [{
-					"label": "20250104-100000F",
-					"annotation": {
-						"multipooler_id": "zone1-multipooler1",
-						"job_id": "20250104-100000.000000_mp-zone1"
-					}
-				}, {
-					"label": "20250104-120000F",
-					"annotation": {
-						"multipooler_id": "zone1-multipooler1",
-						"job_id": "20250104-120000.000000_mp-zone1"
-					}
-				}]
-			}]`,
-			wantBackupID: "20250104-120000F",
-			wantError:    false,
-		},
-		{
-			name:  "No matching backup",
-			jobID: "20250104-180000.000000_mp-zone1",
-			jsonOutput: `[{
-				"backup": [{
-					"label": "20250104-100000F",
-					"annotation": {
-						"multipooler_id": "zone1-multipooler1",
-						"job_id": "20250104-100000.000000_mp-zone1"
-					}
-				}]
-			}]`,
-			wantError:     true,
-			errorContains: "no backup found",
-		},
-		{
-			name:          "No backups at all",
-			jobID:         "20250104-100000.000000_mp-zone1",
-			jsonOutput:    `[{"backup": []}]`,
-			wantError:     true,
-			errorContains: "no backups found",
-		},
-		{
-			name:  "Duplicate matching backups",
-			jobID: "20250104-100000.000000_mp-zone1",
-			jsonOutput: `[{
-				"backup": [{
-					"label": "20250104-100000F",
-					"annotation": {
-						"multipooler_id": "zone1-multipooler1",
-						"job_id": "20250104-100000.000000_mp-zone1"
-					}
-				}, {
-					"label": "20250104-100000F_20250104-110000I",
-					"annotation": {
-						"multipooler_id": "zone1-multipooler1",
-						"job_id": "20250104-100000.000000_mp-zone1"
-					}
-				}]
-			}]`,
-			wantError:     true,
-			errorContains: "found 2 backups",
-		},
-		{
-			name:  "Backup without annotations",
-			jobID: "20250104-100000.000000_mp-zone1",
-			jsonOutput: `[{
-				"backup": [{
-					"label": "20250104-100000F"
-				}]
-			}]`,
-			wantError:     true,
-			errorContains: "no backup found",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Create temp directory for mock pgbackrest binary
-			binDir := t.TempDir()
-
-			// Create mock pgbackrest binary that returns the test JSON
-			mockScript := `#!/bin/bash
-if [[ "$*" == *"info"* ]]; then
-    cat << 'JSONEOF'
-` + tt.jsonOutput + `
-JSONEOF
-    exit 0
-fi
-exit 1
-`
-			pgbackrestPath := binDir + "/pgbackrest"
-			err := exec.Command("sh", "-c", "cat > "+pgbackrestPath+" << 'EOF'\n"+mockScript+"\nEOF").Run()
-			require.NoError(t, err)
-			err = exec.Command("chmod", "+x", pgbackrestPath).Run()
-			require.NoError(t, err)
-
-			// Prepend bin dir to PATH so our mock pgbackrest is found first
-			t.Setenv("PATH", binDir+":/usr/bin:/bin")
-
-			// Use separate directory for pooler data
-			poolerDir := t.TempDir()
-			configPath := setupMockPgBackRestConfig(t, poolerDir)
-			pm := createTestManagerWithBackupLocation(poolerDir, "test-tg", "0", clustermetadatapb.PoolerType_REPLICA, poolerDir)
-			pm.pgBackRestConfigPath = configPath
-
-			ctx := context.Background()
-			backupID, err := pm.findBackupByJobID(ctx, tt.jobID)
-
-			if tt.wantError {
-				require.Error(t, err)
-				if tt.errorContains != "" {
-					assert.Contains(t, err.Error(), tt.errorContains)
-				}
-			} else {
-				require.NoError(t, err)
-				assert.Equal(t, tt.wantBackupID, backupID)
-			}
-		})
-	}
 }
 
 func TestBackup_Validation(t *testing.T) {
@@ -376,6 +232,7 @@ func TestBackup_Validation(t *testing.T) {
 			configPath := setupMockPgBackRestConfig(t, tt.poolerDir)
 			pm := createTestManagerWithBackupLocation(tt.poolerDir, "", "", tt.poolerType, backupLocation)
 			pm.pgBackRestConfigPath = configPath
+			pm.backup.SetConfigPath(configPath)
 
 			// Setup primary info for replica poolers (required for backup)
 			if tt.poolerType == clustermetadatapb.PoolerType_REPLICA {
@@ -429,6 +286,7 @@ func TestGetBackups_Validation(t *testing.T) {
 			configPath := setupMockPgBackRestConfig(t, tt.poolerDir)
 			pm := createTestManager(tt.poolerDir, "", "", clustermetadatapb.PoolerType_REPLICA)
 			pm.pgBackRestConfigPath = configPath
+			pm.backup.SetConfigPath(configPath)
 
 			result, err := pm.GetBackups(ctx, tt.limit)
 
@@ -477,393 +335,6 @@ func TestGetBackups_StatusMapping(t *testing.T) {
 			assert.Equal(t, tt.expectedStatus, status)
 		})
 	}
-}
-
-func TestSafeCombinedOutput(t *testing.T) {
-	tests := []struct {
-		name           string
-		command        string
-		args           []string
-		expectError    bool
-		expectedOutput string
-		outputContains string
-	}{
-		{
-			name:           "Simple echo command",
-			command:        "echo",
-			args:           []string{"hello world"},
-			expectError:    false,
-			expectedOutput: "hello world\n",
-		},
-		{
-			name:           "Command with multiple lines",
-			command:        "printf",
-			args:           []string{"line1\\nline2\\nline3\\n"},
-			expectError:    false,
-			expectedOutput: "line1\nline2\nline3\n",
-		},
-		{
-			name:        "Command that fails",
-			command:     "sh",
-			args:        []string{"-c", "echo 'error message' >&2; exit 1"},
-			expectError: true,
-			// Output should contain the error message from stderr
-			outputContains: "error message",
-		},
-		{
-			name:           "Command with no output",
-			command:        "true",
-			args:           []string{},
-			expectError:    false,
-			expectedOutput: "",
-		},
-		{
-			name:        "Command with both stdout and stderr",
-			command:     "sh",
-			args:        []string{"-c", "echo 'stdout message'; echo 'stderr message' >&2"},
-			expectError: false,
-			// Output should contain both stdout and stderr
-			outputContains: "stdout message",
-		},
-		{
-			name:        "Nonexistent command",
-			command:     "nonexistent-command-12345",
-			args:        []string{},
-			expectError: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			cmd := executil.Command(t.Context(), tt.command, tt.args...)
-			output, err := safeCombinedOutput(cmd)
-
-			if tt.expectError {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-			}
-
-			if tt.expectedOutput != "" {
-				assert.Equal(t, tt.expectedOutput, output)
-			}
-
-			if tt.outputContains != "" {
-				assert.Contains(t, output, tt.outputContains)
-			}
-		})
-	}
-}
-
-func TestSafeCombinedOutput_LargeOutput(t *testing.T) {
-	// Test with large output that could potentially fill the channel buffer
-	// Generate 200 lines (more than the 100-line buffer)
-	t.Run("Large output exceeding channel buffer", func(t *testing.T) {
-		cmd := executil.Command(t.Context(), "sh", "-c", "for i in $(seq 1 200); do echo \"Line $i\"; done")
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		lines := strings.Split(strings.TrimSpace(output), "\n")
-		assert.Equal(t, 200, len(lines), "Should capture all 200 lines")
-		assert.Contains(t, output, "Line 1")
-		assert.Contains(t, output, "Line 200")
-	})
-
-	// Test with very large output (thousands of lines)
-	t.Run("Very large output (1000 lines)", func(t *testing.T) {
-		cmd := executil.Command(t.Context(), "sh", "-c", "for i in $(seq 1 1000); do echo \"Line $i\"; done")
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		lines := strings.Split(strings.TrimSpace(output), "\n")
-		assert.Equal(t, 1000, len(lines), "Should capture all 1000 lines")
-		assert.Contains(t, output, "Line 1")
-		assert.Contains(t, output, "Line 1000")
-	})
-
-	// Test with large output on both stdout and stderr simultaneously
-	t.Run("Large output on both stdout and stderr", func(t *testing.T) {
-		script := `
-		for i in $(seq 1 100); do
-			echo "stdout line $i"
-			echo "stderr line $i" >&2
-		done
-		`
-		cmd := executil.Command(t.Context(), "sh", "-c", script)
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		assert.Contains(t, output, "stdout line 1")
-		assert.Contains(t, output, "stdout line 100")
-		assert.Contains(t, output, "stderr line 1")
-		assert.Contains(t, output, "stderr line 100")
-	})
-}
-
-func TestSafeCombinedOutput_LongLines(t *testing.T) {
-	// Test with very long lines to ensure bufio.Scanner handles them
-	t.Run("Very long single line", func(t *testing.T) {
-		// Generate a long string (10KB)
-		longString := strings.Repeat("a", 10*1024)
-		cmd := executil.Command(t.Context(), "echo", longString)
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		assert.Contains(t, output, longString)
-	})
-
-	// Test with multiple long lines
-	t.Run("Multiple long lines", func(t *testing.T) {
-		// Generate 10 lines of 5KB each
-		script := "for i in $(seq 1 10); do printf '%s\\n' \"$(printf 'x%.0s' {1..5000})\"; done"
-		cmd := executil.Command(t.Context(), "bash", "-c", script)
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		lines := strings.Split(strings.TrimSpace(output), "\n")
-		assert.Equal(t, 10, len(lines))
-		for _, line := range lines {
-			assert.Greater(t, len(line), 4000, "Each line should be at least 4KB")
-		}
-	})
-}
-
-// TestSafeCombinedOutput_LineExceedsScannerLimit verifies that a line
-// longer than bufio.Scanner's default buffer (64 KiB) does not silently
-// truncate the output: safeCombinedOutput surfaces the underlying
-// bufio.ErrTooLong as a wrapped error instead of returning a clean nil.
-// Without the scanner.Err() check inside the reader goroutine, this case
-// previously looked like a successful command with truncated output.
-func TestSafeCombinedOutput_LineExceedsScannerLimit(t *testing.T) {
-	// printf with no newline emits a single token larger than the
-	// default 64 KiB scanner buffer.
-	overlong := strings.Repeat("a", bufio.MaxScanTokenSize+1024)
-	script := "printf %s '" + overlong + "'"
-	cmd := executil.Command(t.Context(), "bash", "-c", script)
-
-	_, err := safeCombinedOutput(cmd)
-	require.Error(t, err, "scanner error on overlong line must be surfaced")
-	assert.True(t, errors.Is(err, bufio.ErrTooLong),
-		"expected wrapped bufio.ErrTooLong, got %v", err)
-}
-
-func TestSafeCombinedOutput_RapidOutput(t *testing.T) {
-	// Test with very rapid output to stress-test the channel and goroutine coordination
-	t.Run("Rapid burst of output", func(t *testing.T) {
-		// Use yes command to generate rapid output, limited by head
-		cmd := executil.Command(t.Context(), "sh", "-c", "yes 'rapid output line' | head -n 500")
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		lines := strings.Split(strings.TrimSpace(output), "\n")
-		assert.Equal(t, 500, len(lines))
-		for _, line := range lines {
-			assert.Equal(t, "rapid output line", line)
-		}
-	})
-}
-
-func TestSafeCombinedOutput_InterleavedOutput(t *testing.T) {
-	// Test with interleaved stdout and stderr to ensure proper handling
-	t.Run("Interleaved stdout and stderr", func(t *testing.T) {
-		script := `
-		for i in $(seq 1 50); do
-			echo "stdout $i"
-			echo "stderr $i" >&2
-		done
-		`
-		cmd := executil.Command(t.Context(), "sh", "-c", script)
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		// Both stdout and stderr should be captured
-		assert.Contains(t, output, "stdout 1")
-		assert.Contains(t, output, "stdout 50")
-		assert.Contains(t, output, "stderr 1")
-		assert.Contains(t, output, "stderr 50")
-	})
-}
-
-func TestSafeCombinedOutput_EmptyStreams(t *testing.T) {
-	t.Run("Only stdout", func(t *testing.T) {
-		cmd := executil.Command(t.Context(), "echo", "only stdout")
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		assert.Equal(t, "only stdout\n", output)
-	})
-
-	t.Run("Only stderr", func(t *testing.T) {
-		cmd := executil.Command(t.Context(), "sh", "-c", "echo 'only stderr' >&2")
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		assert.Contains(t, output, "only stderr")
-	})
-
-	t.Run("Neither stdout nor stderr", func(t *testing.T) {
-		cmd := executil.Command(t.Context(), "true")
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		assert.Empty(t, output)
-	})
-}
-
-func TestSafeCombinedOutput_SlowProducer(t *testing.T) {
-	// Test with a slow producer to ensure goroutines don't deadlock waiting
-	t.Run("Slow output producer", func(t *testing.T) {
-		// Produce output slowly (10 lines with small delays)
-		script := `
-		for i in $(seq 1 10); do
-			echo "line $i"
-			sleep 0.01
-		done
-		`
-		cmd := executil.Command(t.Context(), "sh", "-c", script)
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		lines := strings.Split(strings.TrimSpace(output), "\n")
-		assert.Equal(t, 10, len(lines))
-	})
-}
-
-func TestSafeCombinedOutput_PipeCreationFailure(t *testing.T) {
-	// Test error handling when pipe creation might fail
-	// This is difficult to test directly, but we can test the code path
-	t.Run("Command execution after successful pipe setup", func(t *testing.T) {
-		// This tests that the function properly handles the happy path
-		cmd := executil.Command(t.Context(), "echo", "test")
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		assert.Equal(t, "test\n", output)
-	})
-}
-
-func TestSafeCombinedOutput_StressTest(t *testing.T) {
-	// Stress test: Generate output that exceeds the channel buffer by a large margin
-	// while also mixing stdout and stderr
-	t.Run("Extreme stress test", func(t *testing.T) {
-		// Generate 2000 lines on stdout and 2000 lines on stderr
-		script := `
-		(for i in $(seq 1 2000); do echo "stdout $i"; done) &
-		(for i in $(seq 1 2000); do echo "stderr $i" >&2; done) &
-		wait
-		`
-		cmd := executil.Command(context.Background(), "sh", "-c", script)
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		// Should contain both first and last lines from both streams
-		assert.Contains(t, output, "stdout 1")
-		assert.Contains(t, output, "stdout 2000")
-		assert.Contains(t, output, "stderr 1")
-		assert.Contains(t, output, "stderr 2000")
-
-		// Count approximate number of lines (should be ~4000)
-		lines := strings.Split(strings.TrimSpace(output), "\n")
-		assert.Greater(t, len(lines), 3900, "Should have close to 4000 lines")
-	})
-}
-
-func TestSafeCombinedOutput_BinaryOutput(t *testing.T) {
-	// Test with binary output to ensure it doesn't break the scanner
-	t.Run("Binary-like output", func(t *testing.T) {
-		// Generate output with various characters
-		cmd := executil.Command(context.Background(), "sh", "-c", "printf 'text\\x00with\\x00nulls\\n'")
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		// Should handle the output without crashing
-		assert.NotEmpty(t, output)
-	})
-}
-
-func TestSafeCombinedOutput_CommandNotFound(t *testing.T) {
-	t.Run("Command does not exist", func(t *testing.T) {
-		cmd := executil.Command(context.Background(), "this-command-definitely-does-not-exist-12345")
-		_, err := safeCombinedOutput(cmd)
-
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "failed to start command")
-	})
-}
-
-func TestSafeCombinedOutput_RealWorldScenario(t *testing.T) {
-	// Simulate pgbackrest-like output with mixed stdout/stderr
-	t.Run("Simulate pgbackrest info output", func(t *testing.T) {
-		script := `
-		cat <<'EOF'
-P00   INFO: backup command begin 2.41: --exec-id=12345
-P00   INFO: execute non-exclusive pg_start_backup()
-P00   INFO: backup start archive = 000000010000000000000002
-P00   INFO: new backup label = 20250104-100000F
-P00   INFO: full backup size = 25.3MB
-P00   INFO: backup command end: completed successfully
-EOF
-		`
-		cmd := executil.Command(context.Background(), "sh", "-c", script)
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		assert.Contains(t, output, "new backup label = 20250104-100000F")
-		assert.Contains(t, output, "backup command end: completed successfully")
-	})
-}
-
-func TestSafeCombinedOutput_ConcurrentReads(t *testing.T) {
-	// Verify that concurrent reads from stdout and stderr don't cause issues
-	t.Run("Heavy concurrent output", func(t *testing.T) {
-		script := `
-		# Write to stdout and stderr concurrently as fast as possible
-		(seq 1 1000 | while read i; do echo "OUT$i"; done) &
-		(seq 1 1000 | while read i; do echo "ERR$i" >&2; done) &
-		wait
-		`
-		cmd := executil.Command(context.Background(), "sh", "-c", script)
-		output, err := safeCombinedOutput(cmd)
-
-		require.NoError(t, err)
-		// Both streams should be captured
-		assert.Contains(t, output, "OUT1")
-		assert.Contains(t, output, "ERR1")
-		// Count lines to ensure we got most/all output
-		lineCount := len(strings.Split(strings.TrimSpace(output), "\n"))
-		assert.Greater(t, lineCount, 1900, "Should capture most of the 2000 lines")
-	})
-}
-
-func BenchmarkSafeCombinedOutput(b *testing.B) {
-	b.Run("Small output", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			cmd := executil.Command(context.Background(), "echo", "hello world")
-			_, _ = safeCombinedOutput(cmd)
-		}
-	})
-
-	b.Run("Medium output (100 lines)", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			cmd := executil.Command(context.Background(), "sh", "-c", "for i in $(seq 1 100); do echo \"Line $i\"; done")
-			_, _ = safeCombinedOutput(cmd)
-		}
-	})
-
-	b.Run("Large output (1000 lines)", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			cmd := executil.Command(context.Background(), "sh", "-c", "for i in $(seq 1 1000); do echo \"Line $i\"; done")
-			_, _ = safeCombinedOutput(cmd)
-		}
-	})
-
-	b.Run("Mixed stdout and stderr", func(b *testing.B) {
-		script := "for i in $(seq 1 100); do echo \"stdout $i\"; echo \"stderr $i\" >&2; done"
-		for i := 0; i < b.N; i++ {
-			cmd := executil.Command(context.Background(), "sh", "-c", script)
-			_, _ = safeCombinedOutput(cmd)
-		}
-	})
 }
 
 func TestBackup_ActionLock(t *testing.T) {
@@ -1028,61 +499,6 @@ func TestGetBackupByJobId_Found(t *testing.T) {
 	assert.Equal(t, "20251203-143045.123456_mp-cell-1", resp.Backup.JobId)
 }
 
-func TestPgbackrestConfig(t *testing.T) {
-	t.Run("Success when config path is set", func(t *testing.T) {
-		poolerDir := t.TempDir()
-		configPath := setupMockPgBackRestConfig(t, poolerDir)
-		pm := createTestManagerWithBackupLocation(
-			poolerDir,
-			"test-tg",
-			"0",
-			clustermetadatapb.PoolerType_REPLICA,
-			"/tmp/test-backups",
-		)
-		pm.pgBackRestConfigPath = configPath
-
-		result, err := pm.pgBackRestConfig()
-		require.NoError(t, err)
-		assert.Equal(t, configPath, result)
-	})
-
-	t.Run("Error when config path is not set", func(t *testing.T) {
-		poolerDir := t.TempDir()
-		pm := createTestManagerWithBackupLocation(
-			poolerDir,
-			"test-tg",
-			"0",
-			clustermetadatapb.PoolerType_REPLICA,
-			"/tmp/test-backups",
-		)
-
-		result, err := pm.pgBackRestConfig()
-		require.Error(t, err)
-		assert.Empty(t, result)
-		assert.Contains(t, err.Error(), "not yet generated")
-	})
-
-	t.Run("Repeated calls return same path", func(t *testing.T) {
-		poolerDir := t.TempDir()
-		configPath := setupMockPgBackRestConfig(t, poolerDir)
-		pm := createTestManagerWithBackupLocation(
-			poolerDir,
-			"test-tg",
-			"0",
-			clustermetadatapb.PoolerType_REPLICA,
-			"/tmp/backups",
-		)
-		pm.pgBackRestConfigPath = configPath
-
-		result1, err := pm.pgBackRestConfig()
-		require.NoError(t, err)
-
-		result2, err := pm.pgBackRestConfig()
-		require.NoError(t, err)
-		assert.Equal(t, result1, result2, "should return same path on repeated calls")
-	})
-}
-
 func TestBackup_UsesCLIArgs(t *testing.T) {
 	// Test that Backup() uses CLI args with pgctld-generated config instead of creating temp config
 	ctx := context.Background()
@@ -1145,6 +561,7 @@ pg1-path=/tmp/pg_data
 	// Create test manager with the pooler directory
 	pm := createTestManagerWithBackupLocation(poolerDir, "", "", clustermetadatapb.PoolerType_REPLICA, tmpDir)
 	pm.pgBackRestConfigPath = pgctldConfigPath
+	pm.backup.SetConfigPath(pgctldConfigPath)
 
 	// Setup primary info (required for replica backups)
 	setBackupPrimary(pm, "primary-pooler", "primary.local", 5432)
@@ -1274,12 +691,14 @@ exit 0
 					},
 				}),
 				state:                ManagerStateReady,
-				backupConfig:         backupConfig,
 				actionLock:           actionlock.NewActionLock(),
 				logger:               slog.Default(),
 				pgMonitor:            timer.NewPeriodicRunner(context.TODO(), 10*time.Second),
 				pgBackRestConfigPath: configPath,
 			}
+			pm.backup, _ = backupengine.NewEngine(pm.logger, pm.runLongCommand, pm.record, backupengine.Settings{})
+			pm.backup.SetBackupConfig(backupConfig)
+			pm.backup.SetConfigPath(configPath)
 
 			_, err := pm.Backup(ctx, true, "full", "test-job-id", nil)
 			require.Error(t, err, "backup with empty %s should be rejected", tt.name)
