+++ go/services/multipooler/internal/manager/rpc_verify_backups_test.go
@@ -16,11 +16,7 @@ package manager
 
 import (
 	"context"
-	"fmt"
-	"os"
-	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/require"
 
@@ -30,8 +26,8 @@ import (
 func TestVerifyBackups(t *testing.T) {
 	// VerifyBackups requires a running pgbackrest with a valid stanza.
 	// The actual invocation, including the corrupt-backup negative path, is
-	// tested by integration tests (endtoend/multipooler). This test validates
-	// the precondition checks.
+	// tested by integration tests (endtoend/multipooler) and the backup engine
+	// unit tests. This test validates the manager handler's precondition checks.
 
 	t.Run("fails when not ready", func(t *testing.T) {
 		pm := &MultiPoolerManager{}
@@ -43,106 +39,8 @@ func TestVerifyBackups(t *testing.T) {
 	t.Run("fails when config not yet generated", func(t *testing.T) {
 		// Ready, but topology hasn't been loaded so no config path exists.
 		pm := createTestManager(t.TempDir(), "", "", clustermetadatapb.PoolerType_REPLICA)
-		pm.pgBackRestConfigPath = ""
 		_, err := pm.VerifyBackups(context.Background())
 		require.Error(t, err)
 		require.Contains(t, err.Error(), "pgbackrest config not found")
 	})
-}
-
-// TestVerifyBackups_ExecPaths drives VerifyBackups end-to-end against a stubbed
-// `pgbackrest` binary on PATH. It exercises the three outcomes the production
-// code must distinguish, which the unit-level integration tests cannot reach
-// in-process: a healthy stanza, pgBackRest reporting corruption while still
-// exiting 0 (the case this RPC exists to catch), and pgBackRest exiting non-zero.
-func TestVerifyBackups_ExecPaths(t *testing.T) {
-	tests := []struct {
-		name        string
-		stdout      string
-		exitCode    int
-		wantErr     bool
-		errContains string
-	}{
-		{
-			name:     "healthy stanza succeeds",
-			stdout:   "P00   INFO: stanza: multigres\n    status: ok\n",
-			exitCode: 0,
-			wantErr:  false,
-		},
-		{
-			// pgBackRest exits 0 even when verify finds problems, so the only
-			// signal is the "status: error" line in its output.
-			name:        "corrupt stanza surfaces as error",
-			stdout:      "P00   INFO: invalid checksum 'x.zst'\nP00   INFO: stanza: multigres\n    status: error\n",
-			exitCode:    0,
-			wantErr:     true,
-			errContains: "verify found problems",
-		},
-		{
-			name:        "nonzero exit surfaces as error",
-			stdout:      "P00  ERROR: unable to load info file\n",
-			exitCode:    1,
-			wantErr:     true,
-			errContains: "pgbackrest verify failed",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			binDir := t.TempDir()
-			script := fmt.Sprintf("#!/bin/bash\ncat <<'PGBR_EOF'\n%sPGBR_EOF\nexit %d\n", tt.stdout, tt.exitCode)
-			require.NoError(t, os.WriteFile(filepath.Join(binDir, "pgbackrest"), []byte(script), 0o755))
-			t.Setenv("PATH", binDir+":"+os.Getenv("PATH"))
-
-			poolerDir := t.TempDir()
-			pm := createTestManager(poolerDir, "", "", clustermetadatapb.PoolerType_REPLICA)
-			pm.pgBackRestConfigPath = setupMockPgBackRestConfig(t, poolerDir)
-
-			resp, err := pm.VerifyBackups(context.Background())
-			if tt.wantErr {
-				require.Error(t, err)
-				require.Contains(t, err.Error(), tt.errContains)
-				return
-			}
-			require.NoError(t, err)
-			require.NotNil(t, resp)
-			require.Contains(t, resp.RawOutput, "status: ok")
-			require.GreaterOrEqual(t, resp.Duration, time.Duration(0))
-		})
-	}
-}
-
-// TestVerifyStanzaErrorDetection covers the parser that turns pgBackRest's
-// exit-0-but-reported-error output into a failed RPC. pgBackRest's verify
-// command exits 0 even when it finds a corrupt backup or invalid WAL, so the
-// stanza "status: error" line is the signal we key on.
-func TestVerifyStanzaErrorDetection(t *testing.T) {
-	tests := []struct {
-		name      string
-		output    string
-		wantError bool
-	}{
-		{
-			name:      "healthy stanza",
-			output:    "INFO: stanza: multigres\n    status: ok\n",
-			wantError: false,
-		},
-		{
-			name: "corrupt backup",
-			output: "INFO: invalid checksum '20260605-112558F/pg_data/base/1/2838.zst'\n" +
-				"INFO: stanza: multigres\n    status: error\n" +
-				"      backup: 20260605-112558F, status: invalid, total files checked: 999, total valid files: 998\n" +
-				"        checksum invalid: 1\n",
-			wantError: true,
-		},
-		{
-			name:      "empty output",
-			output:    "",
-			wantError: false,
-		},
-	}
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			require.Equal(t, tc.wantError, verifyStanzaErrorRe.MatchString(tc.output))
-		})
-	}
 }
